requirements.txt
numpy==1.18.1
matplotlib==3.3.1
trajectory_planning_helpers==0.76
# the following is required for minimum time optimization only
casadi==3.5.1
scipy==1.3.3
scikit-learn==0.23.1
# For input_helpers
scikit-image==0.16.2
pyyaml==6.0
pandas==1.3.5
argparse